requirements.txt
requests>=2.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
scipy>=1.10.0
streamlit>=1.28.0
fpdf2>=2.7.0
mercadopago>=2.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
psycopg2-binary>=2.9.0
